+++ ClientApp/src/AnalyticsPage.jsx
@@ -13,9 +13,21 @@ export default function AnalyticsPage() {
     if (!redirectCode) return;
 
     const fetchAnalytics = async () => {
+      const token = localStorage.getItem('token');
+      if (!token) {
+        setError('You must be logged in to view analytics.');
+        setLoading(false);
+        return;
+      }
       try {
         setLoading(true);
-        const res = await fetch(`/api/url/analytics/${redirectCode}`);
+        const res = await fetch(`/api/url/analytics/${redirectCode}`, {
+          method: 'GET',
+          headers: {
+            'Content-Type': 'application/json',
+            'Authorization': `Bearer ${token}` // Ensure token is sent
+          }
+        });
         if (!res.ok) throw new Error('Failed to fetch analytics');
         const data = await res.json();
         setAnalytics(data);
